Execute.c Machine.c OpsList.h SrcTree.c SrcTree.h: Removed some bogus 	production rules wrt dimension names of variables. This in turn 	reduded the instruction set, changed the grammer, scanner and 	translatation phase.
defs.h: Changed all subscripts to be longs so that any integer type can
	be used as a subscript.

Keywords.h: Made QUIT recongition case insensitive.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1759 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/defs.h
@@ -1,6 +1,6 @@
 
 /*
- *      $Id: defs.h,v 1.7 1994-01-13 22:14:25 ethan Exp $
+ *      $Id: defs.h,v 1.8 1994-01-21 02:49:18 ethan Exp $
  */
 /************************************************************************
 *									*
@@ -143,15 +143,15 @@ typedef struct _NclFrame{
 
 typedef struct _NclVectorSelection{
         int n_ind;
-        int *ind;
-	int min;
-	int max;
+        long *ind;
+	long min;
+	long max;
 }NclVectorSelection;
 
 typedef struct _NclSubscriptSelection{
-        int start;
-        int finish;
-        int stride;
+        long start;
+        long finish;
+        long stride;
 }NclSubscriptSelection;
 
 typedef enum {
